+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version latest (0.1.7?, 2019-09-??)
 
+- enh: split out code for the listener design pattern for utilization
 - ...